Makes changes to switch from Canvas to CSS strategy
The intent is for each Cell to have its contents, if any, declare their
CSS class and for that to be of the form typeofthing-n where n is the
size. I will then be able to render each cell contents by just changing
the cell's css class accordingly.

Various changes in the extant code were made in this commit in support
of that.